+++ src/main/java/project/luckybooky/domain/certification/sms/util/SmsCertificationUtil.java
@@ -1,4 +1,4 @@
-package project.luckybooky.domain.certification.util;
+package project.luckybooky.domain.certification.sms.util;
 
 import jakarta.annotation.PostConstruct;
 import net.nurigo.sdk.NurigoApp;
